Move generic list fetching logic outside of route
