+++ apps/web/pages/s/[...slugs]/index.tsx
@@ -1,11 +1,11 @@
-import { useMemo, type FC } from 'react'
+import { type FC } from 'react'
 
 import withApollo from '@island.is/web/graphql/withApollo'
-import { useI18n, withLocale } from '@island.is/web/i18n'
+import { withLocale } from '@island.is/web/i18n'
 import type { LayoutProps } from '@island.is/web/layouts/main'
-import GenericListItemPage, {
-  type GenericListItemPageProps,
-} from '@island.is/web/screens/GenericList/GenericListItem'
+import OrganizationSubPageGenericListItem, {
+  OrganizationSubPageGenericListItemProps,
+} from '@island.is/web/screens/GenericList/OrganizationSubpageGenericListItem'
 import Home, {
   type HomeProps,
 } from '@island.is/web/screens/Organization/Home/Home'
@@ -30,8 +30,6 @@ import SubPage, {
 import { Screen as ScreenType } from '@island.is/web/types'
 import { CustomNextError } from '@island.is/web/units/errors'
 import { getServerSidePropsWrapper } from '@island.is/web/utils/getServerSidePropsWrapper'
-import { useLinkResolver } from '@island.is/web/hooks'
-import { useRouter } from 'next/router'
 
 enum PageType {
   FRONTPAGE = 'frontpage',
@@ -52,61 +50,9 @@ const pageMap: Record<PageType, FC<any>> = {
   [PageType.ALL_EVENTS]: (props) => <OrganizationEventList {...props} />,
   [PageType.NEWS_DETAILS]: (props) => <OrganizationNewsArticle {...props} />,
   [PageType.EVENT_DETAILS]: (props) => <OrganizationEventArticle {...props} />,
-  [PageType.GENERIC_LIST_ITEM]: (props) => {
-    const { organizationPage, subpage } = props.parentProps.componentProps
-    const router = useRouter()
-    const { linkResolver } = useLinkResolver()
-    const backLinkUrl = useMemo(() => {
-      const pathname = new URL(router.asPath, 'https://island.is').pathname
-      return pathname.slice(0, pathname.lastIndexOf('/'))
-    }, [router.asPath])
-    const { activeLocale } = useI18n()
-    return (
-      <SubPage
-        layoutProps={props.parentProps.layoutProps}
-        componentProps={{
-          ...props.parentProps.componentProps,
-          customContent: (
-            <GenericListItemPage
-              item={props.genericListItemProps.item}
-              ogTitle={
-                props.genericListItemProps.item.title &&
-                `${props.genericListItemProps.item.title}${
-                  props.subpage?.title ? ` | ${props.subpage.title}` : ''
-                }`
-              }
-            />
-          ),
-          customBreadcrumbItems: [
-            {
-              title: 'Ísland.is',
-              href: linkResolver('homepage').href,
-            },
-            {
-              title: organizationPage?.title ?? '',
-              href: linkResolver('organizationpage', [
-                organizationPage?.slug ?? '',
-              ]).href,
-            },
-            {
-              title: subpage?.title ?? '',
-              href: backLinkUrl,
-              isTag: true,
-            },
-          ],
-          backLink: {
-            text: activeLocale === 'is' ? 'Til baka' : 'Go back',
-            url: backLinkUrl,
-          },
-          customContentfulIds: [
-            organizationPage?.id,
-            subpage?.id,
-            props.genericListItemProps.item.id,
-          ],
-        }}
-      />
-    )
-  },
+  [PageType.GENERIC_LIST_ITEM]: (props) => (
+    <OrganizationSubPageGenericListItem {...props} />
+  ),
 }
 
 interface Props {
@@ -162,13 +108,7 @@ interface Props {
       }
     | {
         type: PageType.GENERIC_LIST_ITEM
-        props: {
-          parentProps: {
-            layoutProps: LayoutProps
-            componentProps: SubPageProps
-          }
-          genericListItemProps: GenericListItemPageProps
-        }
+        props: OrganizationSubPageGenericListItemProps
       }
 }
 
@@ -290,18 +230,10 @@ Component.getProps = async (context) => {
       }
     }
 
-    // Generic list item
-    const [subPageProps, genericListItemProps] = await Promise.all([
-      SubPage.getProps?.(context),
-      GenericListItemPage.getProps?.(context),
-    ])
     return {
       page: {
         type: PageType.GENERIC_LIST_ITEM,
-        props: {
-          parentProps: subPageProps,
-          genericListItemProps,
-        },
+        props: await OrganizationSubPageGenericListItem.getProps?.(context),
       },
     }
   }
